Use Via for PDF assignments
Use via.hypothes.is, instead of the LTI app's built-in copy of PDF.js
and PDF file caching and serving, to render PDF assignments.

Code related to caching and serving PDF files from the LTI app is
removed.

I've left the static PDF.js code itself in though, so that I can add
in-app proxying of PDF files as an alternative to Via later.